Moved list into subdir, not fully functional yet
