+++ internal/screens/jobs/list/confirmstop.go
@@ -1,4 +1,4 @@
-package jobs
+package list
 
 import (
 	"strings"
